requirements.txt
requests
django==2.2.6
djangorestframework
djangorestframework-simplejwt==4.7.2
django-filter==2.4.0

